requirements.txt
gradio>=4.44.1
ollama>=0.4.0
openai-whisper>=20231117
pydub>=0.25.1
numpy>=1.24.0
soundfile>=0.12.1
scipy>=1.11.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0

# Desktop app dependencies
pyperclip>=1.8.2
pyautogui>=0.9.54
pynput>=1.7.6
sounddevice>=0.4.6
pystray>=0.19.4
Pillow>=10.0.0
